% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcats.R
\name{parcats-shiny}
\alias{parcats-shiny}
\alias{parcatsOutput}
\alias{render_parcats}
\title{Shiny bindings for parcats}
\usage{
parcatsOutput(outputId, width = "100\%", height = "100\%", inline = FALSE)

render_parcats(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{inline, }{logical, Default: FALSE}

\item{expr}{An expression that generates a parcats}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
No return value, called for side effects
}
\description{
Output and render functions for using parcats within Shiny
applications and interactive Rmd documents.
}
